move function superlikeProfile from droppedtags to searchloc
